+++ src/app/app-routing.module.ts
@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
+  { path: '', loadChildren: './auth/auth.module#AuthModule', pathMatch: 'full' },
+  { path: 'content', loadChildren: './content/content.module#ContentModule' },
 ];
 
 @NgModule({
