+++ src/app/app.component.ts
@@ -1,9 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { Auth } from './auth/auth.module';
-import { Store } from '@ngrx/store';
-import * as fromRoot from './reducers';
-import * as category from './actions/categories';
+import { Auth } from './shared/auth-services';
 
 @Component({
   selector: 'app-root',
@@ -13,8 +10,7 @@ import * as category from './actions/categories';
 export class AppComponent {
   constructor(
     public auth: Auth,
-    private router: Router,
-    private store: Store<fromRoot.State>
+    private router: Router
   ) {
     let isLoggedIn: boolean = false;
     this.auth.isLoggedIn.subscribe(newState => {
@@ -25,13 +21,13 @@ export class AppComponent {
     });
   }
 
-  secretPhraseChange(secret: string) {
+  /*secretPhraseChange(secret: string) {
     this.store.dispatch(new category.SecretPhraseChangeAction(secret) as any);
-  }
+  }*/
 
   private logout() {
-    this.store.dispatch(new category.SecretPhraseChangeAction(undefined) as any);
-    this.store.dispatch({ type: 'USER_LOGOUT' });
+    /*this.store.dispatch(new category.SecretPhraseChangeAction(undefined) as any);
+    this.store.dispatch({ type: 'USER_LOGOUT' });*/
     this.router.navigate(['/']);
   }
 }
