+++ src/app/auth/auth.component.html
@@ -1,0 +1,7 @@
+<div fxFlexFill fxLayout="column">
+  <app-header></app-header>
+
+  <main fxFlex fxLayout fxLayoutAlign="center center">
+    <router-outlet></router-outlet>
+  </main>
+</div>
