+++ src/app/auth/auth.module.ts
@@ -1,34 +1,19 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MaterialModule } from '@angular/material';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { ReactiveFormsModule } from '@angular/forms';
 
 import { AuthRoutingModule } from './auth-routing.module';
+import { SharedModule } from '../shared/shared.module';
+
 import { LoginComponent } from './components/login.component';
-import { Auth } from './services/auth.service';
-import { AuthGuard } from './guards/auth-guard';
-import { UnAuthGuard } from './guards/unauth-guard';
+import { AuthComponent } from './auth.component';
 
 @NgModule({
   imports: [
-    CommonModule,
-    MaterialModule,
-    FlexLayoutModule,
-    ReactiveFormsModule,
-    AuthRoutingModule
+    AuthRoutingModule,
+    SharedModule,
   ],
   declarations: [
+    AuthComponent,
     LoginComponent
   ],
-  providers: [
-    Auth,
-    AuthGuard,
-    UnAuthGuard
-  ]
 })
 export class AuthModule { }
-
-export { Auth }
-export { AuthGuard }
-export { UnAuthGuard }
