+++ src/app/auth/components/login.component.html
@@ -13,11 +13,11 @@
       <form novalidate [formGroup]="authForm" (ngSubmit)="login()">
         <div>
           <md-input-container>
-            <input md-input formControlName="username" placeholder="User name" required type="text" [autocomplete]="userNameAutocompleteState" autofocus>
+            <input mdInput formControlName="username" placeholder="User name" required type="text" [autocomplete]="userNameAutocompleteState" autofocus>
           </md-input-container>
 
           <md-input-container>
-            <input md-input formControlName="password" placeholder="Password" required type="password" autocomplete="off">
+            <input mdInput formControlName="password" placeholder="Password" required type="password" autocomplete="off">
           </md-input-container>
         </div>
 
@@ -28,7 +28,7 @@
 
           <div fxLayout fxLayoutAlign="start center" fxLayoutGap="5px">
             <md-input-container>
-              <input md-input formControlName="exp_delta_amount" placeholder="Type in amount" type="number">
+              <input mdInput formControlName="exp_delta_amount" placeholder="Type in amount" type="number">
             </md-input-container>
 
             <md-select name="items" formControlName="exp_delta_multiplier" placeholder="Select interval">
