--- src/app/auth/guards/auth-guard.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { AuthGuard } from './auth-guard';
-
-describe('AuthGuardService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AuthGuard]
-    });
-  });
-
-  it('should ...', inject([AuthGuard], (service: AuthGuard) => {
-    expect(service).toBeTruthy();
-  }));
-});
