--- src/app/auth/services/auth.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { Auth } from './auth.service';
-
-describe('AuthService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [Auth]
-    });
-  });
-
-  it('should ...', inject([Auth], (service: Auth) => {
-    expect(service).toBeTruthy();
-  }));
-});
