+++ src/app/content/content-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { ContentComponent } from './content.component';
-import { AuthGuard } from '../auth/auth.module';
+import { AuthGuard } from '../shared/auth-services';
 
 const routes: Routes = [
-  { path: 'content', component: ContentComponent, canActivate: [AuthGuard] }
+  { path: '', component: ContentComponent, canActivate: [AuthGuard] }
 ];
 
 @NgModule({
