+++ src/app/shared/app-header/app.header.component.html
@@ -3,7 +3,7 @@
     <span>Open Pass Phrase</span>
 
     <md-input-container *ngIf="loggedIn | async" dividerColor="accent" fxFlex="265px">
-      <input md-input [formControl]="secretPhrase" autocomplete="off" type="password" placeholder="secret phrase to decode data" #secretInput>
+      <input mdInput [formControl]="secretPhrase" autocomplete="off" type="password" placeholder="secret phrase to decode data" #secretInput>
       <md-hint *ngIf="secretPhrase.dirty && !secretPhrase.valid">Must be at least 6 chars</md-hint>
     </md-input-container>
