+++ src/app/shared/auth-services/auth-guard.ts
@@ -4,20 +4,21 @@ import 'rxjs/add/operator/take';
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { CanActivate } from '@angular/router';
-import { Auth } from '../services/auth.service';
+import { Auth } from './auth.service';
+import { Observable } from 'rxjs/Observable';
+import { Subscriber } from 'rxjs/Subscriber';
 
 @Injectable()
-export class UnAuthGuard implements CanActivate {
+export class AuthGuard implements CanActivate {
 
   constructor(private auth: Auth, private router: Router) { }
 
   canActivate() {
     return this.auth.isLoggedIn
       .take(1)
-      .map(isLoggedIn => !isLoggedIn)
-      .do(isNotLoggedIn => {
-        if (!isNotLoggedIn) {
-          this.router.navigate(['content']);
+      .do(isLoggedIn => {
+        if (!isLoggedIn) {
+          this.router.navigate(['/']);
         }
       });
   }
