+++ src/app/shared/auth-services/unauth-guard.ts
@@ -4,21 +4,20 @@ import 'rxjs/add/operator/take';
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { CanActivate } from '@angular/router';
-import { Auth } from '../services/auth.service';
-import { Observable } from 'rxjs/Observable';
-import { Subscriber } from 'rxjs/Subscriber';
+import { Auth } from './auth.service';
 
 @Injectable()
-export class AuthGuard implements CanActivate {
+export class UnAuthGuard implements CanActivate {
 
   constructor(private auth: Auth, private router: Router) { }
 
   canActivate() {
     return this.auth.isLoggedIn
       .take(1)
-      .do(isLoggedIn => {
-        if (!isLoggedIn) {
-          this.router.navigate(['/']);
+      .map(isLoggedIn => !isLoggedIn)
+      .do(isNotLoggedIn => {
+        if (!isNotLoggedIn) {
+          this.router.navigate(['content']);
         }
       });
   }
