+++ src/app/shared/shared.module.ts
@@ -5,25 +5,30 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MaterialModule } from '@angular/material';
 
 import { AuthModule } from '../auth/auth.module';
-import { AppHeaderComponent } from '../app-header/app.header.component';
+import { AppHeaderComponent } from './app-header/app.header.component';
+import { Auth, AuthGuard, UnAuthGuard } from './auth-services'
 
 @NgModule({
   imports: [
     CommonModule,
     FlexLayoutModule,
-    AuthModule,
     ReactiveFormsModule,
     MaterialModule,
   ],
   exports: [
     CommonModule,
     FlexLayoutModule,
-    AuthModule,
     ReactiveFormsModule,
     MaterialModule,
+    AppHeaderComponent,
   ],
   declarations: [
     AppHeaderComponent,
-  ]
+  ],
+  providers: [
+    Auth,
+    AuthGuard,
+    UnAuthGuard,
+  ],
 })
 export class SharedModule { }
